Refactor: user 관련 타입 분리
